+++ back/src/test/resources/application-test.properties
@@ -1,22 +1,22 @@
+# Database Configuration
 spring.datasource.url=jdbc:h2:mem:testdb;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=false
 spring.datasource.driverClassName=org.h2.Driver
-spring.datasource.username=sa
-spring.datasource.password=
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 spring.jpa.hibernate.ddl-auto=none
-spring.jpa.show-sql=true
+
+# Disable SQL logging
+spring.jpa.show-sql=false
+spring.jpa.properties.hibernate.format_sql=false
+
+# Data initialization
 spring.sql.init.mode=always
 spring.sql.init.schema-locations=classpath:schema-test.sql
 spring.sql.init.data-locations=classpath:data-test.sql
 
-spring.h2.console.enabled=true
+# Disable H2 Console for tests
+spring.h2.console.enabled=false
 spring.h2.console.path=/h2-console
-spring.jpa.properties.hibernate.format_sql=true
 
+# JWT Configuration
 oc.app.jwtSecret=openclassrooms
 oc.app.jwtExpirationMs=86400000
-
-logging.level.org.hibernate.SQL=DEBUG
-logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
-logging.level.org.springframework=DEBUG
-logging.level.org.hibernate=DEBUG
